+++ index.html
@@ -231,7 +231,6 @@
   <template id="messages">
     <div class="img-upload__message  img-upload__message--loading">Загружаем...</div>
   </template>
-  <script src="./js/main.js" defer></script>
-  <script src="./js/functions.js"></script>
+  <script src="./js/main.js" defer type="module"></script>
 </body>
 </html>
